pyproject.toml
[project]
name = "confluence-ragbot"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "chromadb>=1.5.1",
    "llama-index-core>=0.14.15",
    "llama-index-embeddings-huggingface>=0.6.1",
    "llama-index-llms-openai>=0.6.21",
    "llama-index-readers-confluence>=0.6.1",
    "llama-index-readers-file>=0.5.6",
    "llama-index-retrievers-bm25>=0.6.5",
    "llama-index-vector-stores-chroma>=0.5.5",
    "python-dotenv>=1.2.1",
    "python-telegram-bot>=22.6",
]
